fixed #1712: Adding default dpi for marmalade and mac since i don't know how to get DPI for these platforms.

+++ cocos2dx/platform/CCDevice.h
@@ -10,7 +10,11 @@ class CC_DLL CCDevice
 private:
     CCDevice();
 public:
-    static float getDPI();
+    /**
+     *  Gets the DPI of device
+     *  @return The DPI of device.
+     */
+    static int getDPI();
 };
